Move EolympBasecampProblemParser to correct directory
